loadContext: shared code to load alerts, featureFlags, settings, user
